+++ apps/foh/src/app.routes.tsx
@@ -1,41 +1,55 @@
-import { MakeGenerics, Navigate, Route } from '@tanstack/react-location';
-
-import { Championship } from 'lib';
-
-import { SplashScreen } from 'ui';
-import Layout, { rootLoader } from '@/app/layout';
+import Layout from '@/app/layout';
 
 import ChampionshipPage from '@/app/championship/page';
 import PlayersPage from '@/app/players/page';
 import RankingPage from '@/app/ranking/page';
 import MatchesPage from '@/app/matches/page';
+import { redirect, RouteObject } from 'react-router-dom';
+import { Championship, collection, filter, orderByDesc } from 'lib';
+import { queryClient } from './app';
 
-export type LocationGenerics = MakeGenerics<{
-  LoaderData: {
-    championships: Championship[];
-    currentChampionship: Championship | null;
-    selectedChampionship: Championship | null;
-  };
-  Params: {
-    championshipId: string;
-  };
-}>;
-
-export const appRoutes: Route<LocationGenerics>[] = [
+export const appRoutes: RouteObject[] = [
   {
     path: '/',
-    element: <Navigate to="/turnier" />,
+    loader: () => redirect('/turnier'),
   },
   {
+    id: 'root',
     path: ':championshipId',
     element: <Layout />,
-    loader: rootLoader,
-    pendingElement: async () => <SplashScreen />,
+    loader: async ({ params: { championshipId } }) => {
+      const championships = await queryClient.fetchQuery(
+        ['championships'],
+        collection<Championship>(
+          'championships',
+          filter('published', '==', true),
+          orderByDesc('nr')
+        ).get
+      );
+      const currentChampionship =
+        championshipId === 'turnier'
+          ? championships?.at(0) || null
+          : championships?.find((c) => c.id === championshipId);
+
+      return { championships, currentChampionship };
+    },
     children: [
-      { path: '/', element: <ChampionshipPage /> },
-      { path: 'tabelle', element: <RankingPage /> },
-      { path: 'spieler', element: <PlayersPage /> },
-      { path: 'spiele', element: <MatchesPage /> },
+      { index: true, handle: { childPath: '' }, element: <ChampionshipPage /> },
+      {
+        path: 'tabelle',
+        handle: { childPath: '/tabelle' },
+        element: <RankingPage />,
+      },
+      {
+        path: 'spieler',
+        handle: { childPath: '/spieler' },
+        element: <PlayersPage />,
+      },
+      {
+        path: 'spiele',
+        handle: { childPath: '/spiele' },
+        element: <MatchesPage />,
+      },
     ],
   },
 ];