+++ apps/foh/src/app.tsx
@@ -1,10 +1,20 @@
-import { ReactLocation, Router } from '@tanstack/react-location';
-import { appRoutes, LocationGenerics } from './app.routes';
+// React Query
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+export const queryClient = new QueryClient();
 
-const location = new ReactLocation<LocationGenerics>();
+// React Router
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { appRoutes } from './app.routes';
+const router = createBrowserRouter(appRoutes);
 
 function App() {
-  return <Router location={location} routes={appRoutes} />;
+  return (
+    <QueryClientProvider client={queryClient}>
+      <RouterProvider router={router} />
+      <ReactQueryDevtools initialIsOpen={false} />
+    </QueryClientProvider>
+  );
 }
 
 export default App;