+++ apps/foh/src/env.d.ts
@@ -1,1 +1,13 @@
 /// <reference types="vite/client" />
+
+import { Params } from 'react-router-dom';
+
+declare module 'react-router-dom' {
+  function useMatches(): {
+    id: string;
+    pathname: string;
+    params: Params<string>;
+    data: unknown;
+    handle?: { childPath?: string };
+  }[];
+}