+++ apps/foh/src/hooks/use-championships.ts
@@ -1,0 +1,9 @@
+import { Championship } from 'lib';
+import { useRouteLoaderData } from 'react-router-dom';
+
+export function useChampionships() {
+  const { championships } = useRouteLoaderData('root') as {
+    championships: Championship[];
+  };
+  return championships;
+}